requirements.txt
pandas>=1.5.0
antlr4-python3-runtime>=4.13.1
anytree>=2.8.0 